+++ ParkFlow/src/main/java/com/example/parkflow/Service/Impl/SensorServiceImpl.java
@@ -20,6 +20,7 @@ import org.springframework.stereotype.Service;
 import java.math.BigDecimal;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 import java.time.LocalDateTime;
@@ -37,6 +38,25 @@ public class SensorServiceImpl implements SensorService {
         this.userRepository = userRepository;
         this.reservationRepository = reservationRepository;
         this.hubRepository = hubRepository;
+    }
+
+    public boolean isSensorOwnedByUser(Long sensorId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Object principal = authentication.getPrincipal();
+        User user = userRepository.findByEmail(principal.toString()).orElse(null);
+        try {
+            if (user != null) {
+                Optional<Sensor> sensorOptional = sensorRepository.findById(sensorId);
+                if (sensorOptional.isPresent()) {
+                    Sensor sensor = sensorOptional.get();
+                    System.out.println(sensor.getOwner());
+                    return sensor.getOwner().equals(user);
+                }
+            }
+            return false;
+        } catch (Exception e) {
+            return false;
+        }
     }
     private void authorizeCustomerOrAdmin() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -66,10 +86,11 @@ public class SensorServiceImpl implements SensorService {
             throw new ResponseException("Sensor not found.", HttpStatus.NOT_FOUND);
         }
     }
+
     @Override
-    public Sensor create(double latitude, double longitude, Address address) {
+    public Sensor create(double latitude, double longitude, Address address, Boolean isPrivate) {
         authorizeCustomerOrAdmin();
-        var sensor = new Sensor(latitude, longitude, address);
+        var sensor = new Sensor(latitude, longitude, address, isPrivate);
         return sensorRepository.save(sensor);
     }
 
@@ -86,17 +107,21 @@ public class SensorServiceImpl implements SensorService {
     }
 
     @Override
-    public Sensor update(double latitude, double longitude, Address address, Long id) {
+    public Sensor update(double latitude, double longitude, Address address, Long id, Long userId) {
         authorizeCustomerOrAdmin();
-        return sensorRepository.findById(id)
-                .map(sensor -> {
-                    sensor.setLatitude(latitude);
-                    sensor.setLongitude(longitude);
-                    sensor.setAddress(address);
-                    sensor.setUpdatedAtTimestamp(LocalDateTime.now());
-                    return sensorRepository.save(sensor);
-                })
-                .orElse(null);
+        var userRole = userRepository.findById(userId).orElse(null).getAuthorities().stream().findFirst().orElse(null);
+        if(isSensorOwnedByUser(id) || userRole.getAuthority().equals("ADMIN")) {
+            return sensorRepository.findById(id)
+                    .map(sensor -> {
+                        sensor.setLatitude(latitude);
+                        sensor.setLongitude(longitude);
+                        sensor.setAddress(address);
+                        sensor.setUpdatedAtTimestamp(LocalDateTime.now());
+                        return sensorRepository.save(sensor);
+                    })
+                    .orElse(null);
+        }
+        throw new ResponseException("Access denied. You do not own this sensor.", HttpStatus.FORBIDDEN);
     }
 
     @Override
@@ -105,14 +130,20 @@ public class SensorServiceImpl implements SensorService {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id, Long userId) {
         authorizeCustomerOrAdmin();
-        sensorRepository.deleteById(id);
+        var userRole = userRepository.findById(userId).orElse(null).getAuthorities().stream().findFirst().orElse(null);
+        if(isSensorOwnedByUser(id) || userRole.getAuthority().equals("ADMIN")) {
+            sensorRepository.deleteById(id);
+        } else
+            throw new ResponseException("Access denied. You do not own this sensor.", HttpStatus.FORBIDDEN);
     }
 
     @Override
     public List<Sensor> getClosest(double myLatitude, double myLongitude, int number) {
         List<Sensor> sensorList = sensorRepository.findAll();
+        sensorList.removeIf(Sensor::getIsPrivate);
+        sensorList.removeIf(sensor -> sensor.getAddress() == null);
         sensorList.sort(Comparator.comparingDouble(sensor ->
                 calculateDistance(sensor.getLatitude(), sensor.getLongitude(), myLatitude, myLongitude)));
         return sensorList.subList(0, Math.min(number, sensorList.size()));
@@ -132,15 +163,28 @@ public class SensorServiceImpl implements SensorService {
     @Override
     public Sensor reserveSensor(Long sensorId, Long userId) {
         Optional<Sensor> sensorOptional = sensorRepository.findById(sensorId);
-        if (sensorOptional.isPresent()) {
+        Optional<User> userOptional = userRepository.findById(userId);
+        if (sensorOptional.isPresent() && userOptional.isPresent()) {
             Sensor sensor = sensorOptional.get();
-            if (!sensor.isAvailable()) {
+            User user = userOptional.get();
+
+            if (!sensor.isAvailable() || user.getReservedSensorId() != null) {
                 return null;
             }
+
+            System.out.println("Sensor is private: " + sensor.getIsPrivate() + " and user is owner: " + isSensorOwnedByUser(sensorId));
+            if (sensor.getIsPrivate() && !isSensorOwnedByUser(sensorId)) {
+                return null;
+            }
+
             sensor.setReservationStartTimestamp(LocalDateTime.now());
             sensor.setReservedByUserId(userId);
             sensor.setAvailable(false);
             sensorRepository.save(sensor);
+
+            user.setReservedSensorId(sensorId);
+            userRepository.save(user);
+
             return sensor;
         }
         return null;
@@ -149,8 +193,11 @@ public class SensorServiceImpl implements SensorService {
     @Override
     public Reservation endReservation(Long sensorId, Long userId, String paymentMethod) {
         Optional<Sensor> sensorOptional = sensorRepository.findById(sensorId);
-        if (sensorOptional.isPresent()) {
+        Optional<User> userOptional = userRepository.findById(userId);
+        if (sensorOptional.isPresent() && userOptional.isPresent()) {
             Sensor sensor = sensorOptional.get();
+            User user = userOptional.get();
+
             if (sensor.isAvailable()) {
                 return null;
             }
@@ -175,6 +222,9 @@ public class SensorServiceImpl implements SensorService {
             sensor.setLifted(true);
             sensorRepository.save(sensor);
 
+            user.setReservedSensorId(null);
+            userRepository.save(user);
+
             return reservation;
         }
         return null;
@@ -192,12 +242,12 @@ public class SensorServiceImpl implements SensorService {
     }
 
     @Override
-    public boolean updateSensorAvailability(Long sensorId, Boolean available) {
+    public boolean updateSensorPirvateState(Long sensorId, Boolean available) {
         authorizeCustomerOrAdmin();
         Optional<Sensor> sensorOptional = sensorRepository.findById(sensorId);
-        if (sensorOptional.isPresent()) {
+        if (sensorOptional.isPresent() && isSensorOwnedByUser(sensorId)) {
             Sensor sensor = sensorOptional.get();
-            sensor.setAvailable(available);
+            sensor.setIsPrivate(available);
             sensorRepository.save(sensor);
             return true;
         } else {
@@ -218,6 +268,7 @@ public class SensorServiceImpl implements SensorService {
             }
             if (sensor.isLifted()) {
                 sensor.setLifted(false);
+                sensor.setUpdatedAtTimestamp(LocalDateTime.now());
                 sensorRepository.save(sensor);
                 return sensor;
             }
